actions_refactor split actions.py into folders, grouped by usage documented and annotated all actions
