+++ src/action/base.py
@@ -1,0 +1,27 @@
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from engine import Engine
+    from entity import Actor
+
+
+class Action:
+    """Generic Action"""
+    
+    def __init__(self, entity: Actor):
+        self.entity = entity
+    
+    @property
+    def engine(self) -> Engine:
+        """Return engine for this action"""
+        return self.entity.parent.engine
+    
+    def perform(self) -> bool:
+        """Perform this action with the objects needed to determine its scope.
+        `self.engine` is the scope this action is being performed in.
+        `self.entity` is the object performing the action.
+        This method must be overridden by Action subclasses.
+        """
+        raise NotImplementedError()
