+++ src/action/quit.py
@@ -1,0 +1,8 @@
+from action.base import Action
+
+
+class ActionQuit(Action):
+    """Action that quits the game"""
+    
+    def perform(self) -> None:
+        raise SystemExit()
